% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFSW.R
\docType{data}
\name{dataFSW}
\alias{dataFSW}
\title{The data from the study of Polish upper-secondary schools students.}
\format{
data frame: 3796 obs. of 54 variables
}
\description{
It was conducted firstly in the same time as PISA 2009 study, with use of the
same cognitive tests and questionnaires as in PISA 2009, but on a different
group of students: first grade students of upper-secondary schools (in Poland
most of the students in a regular PISA sample attends lower-secondary
schools). The students who participated in the first wave of the study were
followed in the 2nd grade of upper-secondary school within the research
program \emph{Our further study and work} (\emph{Nasza Dalsza Nauka i Praca}).
Both studies were conducted by the Institute of Philosophy and Sociology
Polish Academy of Sciences.
}
\details{
\strong{The original data was changed a little, to better fit the purpose of
the game.}
}
\keyword{datasets}
