% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnglishLetterFrequency.R
\docType{data}
\name{EnglishLetterFrequency}
\alias{EnglishLetterFrequency}
\title{The vector of letter frequencies in English.}
\format{
a named vector with 26 elements.
}
\usage{
data(EnglishLetterFrequency)
}
\description{
The vector with frequencies of 26 English letters.
It is sorted by the frequency of usage.
May be used to refine the transliteraton.
}
\keyword{datasets}
