% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top1000passwords.R
\docType{data}
\name{top1000passwords}
\alias{top1000passwords}
\title{The vector of 1000 most popular passwords.}
\format{
a character vector with 1000 elements.
}
\description{
The character vector of 1000 most commonly used passwords.
It is sorted by the frequency of password's usage. First passwords in the vector are the most frequently used.
}
\keyword{datasets}
