% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top100commonWords.R
\docType{data}
\name{top100commonWords}
\alias{top100commonWords}
\title{The vector of 100 most common words in English.}
\format{a character vector with 100 elements.}
\usage{
data(top100commonWords)
}
\description{
The character vector of 100 most commonly used words in English.
It is sorted by the frequency of usage.
May be used to refine the transliteraton.
}
\keyword{datasets}

