% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varLabels.R
\docType{data}
\name{varLabels}
\alias{varLabels}
\title{The data frame containng labels of the variables from \code{dataDNiP} and
\code{DNiP} datasets.}
\format{
data frame: 54 obs. of 2 variables
}
\description{
The data frame containng labels of the variables from \code{dataDNiP} and
\code{DNiP} datasets.
}
\keyword{datasets}
