% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiquotes.R
\docType{data}
\name{wikiquotes}
\alias{wikiquotes}
\title{List with quotes in 18 languages.}
\format{a named list with 18 elements.}
\usage{
data(wikiquotes)
}
\description{
The named list with 18 languages. Based on <https://wikiquote.org/>.
}
\keyword{datasets}

