% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequon.R
\name{frequon}
\alias{frequon}
\title{The Frequon (Frequency Analysis) Game}
\usage{
frequon(...)
}
\arguments{
\item{...}{\code{frequon} function is called by different arguments, which vary depending
on a problem that Bit is trying to solve. See \code{Details} in order to learn more about the list of possible arguments.}
}
\description{
The \code{frequon} function is used for solving problems in the data-based game ,,The Frequon Game''.
}
\details{
Every time when some additional hints are needed one should add
\code{hint=TRUE} argument to the \code{frequon} function.

In this game you are in contact with a group of people that are going to stop terrorists.
You can communicate with them through \code{frequon} function.

In each call add \code{subject} parameter that will indicate which message you are answering.
Add \code{content} parameter. It's value should match the request.

,,The Frequon Game'' is a free of charge, educational project of the SmarterPoland.pl Foundation.
}
\examples{
frequon()
frequon(hint=TRUE)
}
\author{
Katarzyna Fak - the idea and the implementation,
Przemyslaw Biecek - comments and the integration with this package.
}

